+++ tests/unit/conftest.py
@@ -1,10 +1,14 @@
+import asyncio
 import logging
 import multiprocessing as mp
 import sys
 import tempfile
 import time
 from contextlib import asynccontextmanager
-from multiprocessing import Value, queues
+from multiprocessing import (
+    Value,
+    queues,
+)
 from typing import Tuple
 
 import numpy as np
@@ -13,8 +17,11 @@ import pytest
 from aqueduct.flow import Flow
 from aqueduct.handler import BaseTaskHandler
 from aqueduct.logger import LOGGER_NAME
+from aqueduct.multiprocessing import ProcessContext
 from aqueduct.shm import NPArraySharedData
 from aqueduct.task import BaseTask
+
+pytest_plugins = 'aiohttp.pytest_plugin'
 
 # необходимо для корректной работы тестов на MacOS
 mp.set_start_method('fork')
@@ -87,6 +94,14 @@ class MultiPlatformQueue(queues.Queue):
 
 if sys.platform == 'darwin':
     queues.Queue = MultiPlatformQueue
+
+
+async def stop_flow(flow: Flow) -> Tuple[mp.process.BaseProcess, BaseTaskHandler]:
+    handler, context = next(iter(flow._contexts.items()))  # type: BaseTaskHandler, ProcessContext
+    process = context.processes[0]
+    process.terminate()
+    await asyncio.sleep(1)
+    return process, handler
 
 
 @pytest.fixture
@@ -172,12 +187,12 @@ async def run_flow(flow: Flow):
 
 
 @pytest.fixture
-async def simple_flow(sleep_handlers: Tuple[SleepHandler, ...]) -> Flow:
+async def simple_flow(loop, sleep_handlers: Tuple[SleepHandler, ...]) -> Flow:
     async with run_flow(Flow(*sleep_handlers)) as flow:
         yield flow
 
 
 @pytest.fixture
-async def slow_simple_flow(slow_sleep_handlers: Tuple[SleepHandler, ...]) -> Flow:
+async def slow_simple_flow(loop, slow_sleep_handlers: Tuple[SleepHandler, ...]) -> Flow:
     async with run_flow(Flow(*slow_sleep_handlers)) as flow:
         yield flow