+++ tests/unit/metrics/test_collect.py
@@ -1,8 +1,4 @@
-import pytest
-
-
 class TestCollector:
-    @pytest.mark.asyncio
     async def test_collect_qsize(self, simple_flow):
         qsize_metrics = simple_flow._metrics_manager.exporter.target.queue_sizes.items
         assert len(qsize_metrics) == 4